Moved factories into requests folder
