+++ rmf_task/include/rmf_task/requests/ChargeBatteryFactory.hpp
